backend/containers/django/requirements.txt
Django
psycopg2
psycopg2-binary
gunicorn
djangorestframework
djangorestframework-simplejwt
djangorestframework-gis
django-cors-headers
django-oauth-toolkit
django-environ
djoser
